pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trade"
version = "0.1.0"
description = "Multi-user stock trading CLI — Angel One + AI natural language"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "plotext>=5.2.0",
    "python-dotenv>=1.0.0",
    "pyotp>=2.9.0",
    "smartapi-python>=1.3.4",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "websocket-client>=1.6.0",
    "yfinance>=0.2.0",
    "colorama>=0.4.6",
]

[project.scripts]
trade = "trade.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["trade"]
